# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
      totp:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterReq:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  files.File:
    properties:
      id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
  response.ListResponse:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      details: {}
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: you@example.com
    name: Deva
  description: GoNotes — Notes API with Gin, GORM, OAuth, TOTP, GraphQL (read).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoNotes API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email + password
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register user
      tags:
      - auth
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file
      parameters:
      - description: file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/files.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload file
      tags:
      - files
  /notes:
    get:
      consumes:
      - application/json
      description: Get paginated notes
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListResponse'
      security:
      - ApiKeyAuth: []
      summary: List notes
      tags:
      - notes
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"